+++ app/mailers/analysis_mailer.rb
@@ -6,6 +6,7 @@ class AnalysisMailer < ApplicationMailer
     @analysis = analysis
     subject = is_success ? '分析が完了しました' : '分析中にエラーが発生しました'
     template_name = is_success ? 'success' : 'failer'
+    attachments['training_data.yml'] = File.read(File.join(Rails.root, "results/analysis_#{analysis.id}.yml"))
     mail(:to => 'Leonis.0813@gmail.com', :subject => subject, :template_name => template_name)
   end
 end